ZP Projekat/requirements.txt
cryptography==40.0.2
crypto==1.4.1
pycryptodome==3.18.0
pycryptodomex==3.18.0
wheel==0.40.0
pycparser==2.21

# pip install -r requirements.txt
